requirements.txt
playwright==1.36.0
pandas==2.2.0
sqlalchemy==2.0.19
python-dotenv==1.0.0
apache-airflow==2.6.3
requests==2.31.0
beautifulsoup4==4.12.2